pyproject.toml
[project]
name = "rager"
version = "0.1.0"
description = ""
authors = [
    {name = "noemacee",email = "100673718+noemacee@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.14"

dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "langchain (>=0.3.20,<0.4.0)",
    "openai (>=1.66.3,<2.0.0)",
    "pinecone-client (>=6.0.0,<7.0.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "langchain-community (>=0.3.19,<0.4.0)",
    "accelerate (>=1.5.2,<2.0.0)",
    "langchain-huggingface (>=0.1.2,<0.2.0)",
    "flask (>=3.1.0,<4.0.0)",
    "langchain-core (>=0.3.45,<0.4.0)",
    "langchain-openai (>=0.3.8,<0.4.0)",
    "langchain-pinecone (>=0.2.3,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
